requirements.txt
cloudinary==1.26.0
dj-database-url==0.5.0
Django==3.1.13
django-cloudinary-storage==0.3.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
Pillow==8.1.2
psycopg2==2.9.1
whitenoise==5.3.0
slackweb==1.0.5
slackclient==2.9.3